Change _swift_fmodf to return float
Hoping this fixes the armv7 test failures.

Swift SVN r19811

+++ stdlib/runtime/Stubs.cpp
@@ -117,7 +117,7 @@ extern "C" uint64_t swift_doubleToString(char *Buffer, size_t BufferLength,
 }
 
 
-extern "C" double _swift_fmodf(float lhs, float rhs) {
+extern "C" float _swift_fmodf(float lhs, float rhs) {
     return fmodf(lhs, rhs);
 }
